refactor: move the gallery component to a separate folder.
