+++ src/app/[locale]/components/gallery/ImageGallery.tsx
